requirements.txt
pandas
gnews
python-dotenv
sqlalchemy
psycopg2-binary
google-genai